separate Node.js and web browser examples
