--- example/browser.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>debug()</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-    <script src="../dist/debug.js"></script>
-    <script>
-      // type debug.enable('*') in
-      // the console and refresh :)
-
-      var a = debug('worker:a');
-      var b = debug('worker:b');
-
-      setInterval(function(){
-        a('doing some work');
-      }, 1000);
-
-      setInterval(function(){
-        b('doing some work');
-      }, 1200);
-    </script>
-  </head>
-  <body>
-
-  </body>
-</html>
