--- example/wildcards.js
@@ -1,10 +1,0 @@
-
-var debug = {
-  foo: require('../')('test:foo'),
-  bar: require('../')('test:bar'),
-  baz: require('../')('test:baz')
-};
-
-debug.foo('foo')
-debug.bar('bar')
-debug.baz('baz')
