+++ examples/browser/colors.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>debug()</title>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+    <script src="../../dist/debug.js"></script>
+    <script>
+      debug.enable('*')
+
+      for (var i=0; i < debug.colors.length; i++) {
+        debug('example:' + i)('The color is %o', debug.colors[i])
+      }
+    </script>
+    <style type="text/css">
+      body {
+        padding-top: 100px;
+        text-align: center;
+      }
+    </style>
+  </head>
+  <body>
+    Open your
+    <strong>Web Inspector</strong>
+    to see the <code>debug</code> output
+  </body>
+</html>
